+++ jobbergate-api/CHANGELOG.md
@@ -8,6 +8,7 @@ This file keeps track of all notable changes to jobbergate-api
 ## 5.2.0a1 -- 2024-05-24
 - Enabled template and job-script files to be renamed on upsert routes [PENG-2070]
 - Added a response model for PUT on `/job-scripts/{id}/upload/{file_type}`
+- Improve performance on Automatically clean up unused job scripts [ASP-5186]
 
 ## 5.2.0a0 -- 2024-04-29